add game_teams manager and tests
