+++ spec/spec_helper.rb
@@ -17,6 +17,7 @@ require 'simplecov'
 SimpleCov.start
 require './lib/games_manager'
 require './lib/teams_manager'
+require './lib/game_teams_manager'
 
 RSpec.configure do |config|
   # rspec-expectations config goes here. You can use an alternate
